Fix savegame-compatibility of scripts, increase BUILD_NUMBER
"Fix "t->c->value.argSize == func->parmTotal" Assertion in Scripts, #303"
had broken old savegames because the script checksum
(idProgram::CalculateChecksum()) changed, see #344.
This is fixed now, also the BUILD_NUMBER is increased so old savegames
can be identified for a workaround.

Don't use this commit without the "Add information about dhewm3 build
 to savegames" commit, which further modifies the savegame format
(for the new BUILD_NUMBER 1305). In dhewm3-sdk this commit is directly
before that commit, but unfortunately I fucked up when cherry-picking
the "newer" commits onto librecoop, missing a few of them, so this one
and the previous one and the next three commits are out of order